studiogrid/requirements.txt
temporalio>=1.7.0
pyyaml>=6.0.1
